+++ backend/coins/btc/db/transactionsdb/transactionsdb.go
@@ -26,7 +26,9 @@ import (
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/coins/btc/transactions"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/coins/btc/types"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/coins/btc/util"
+	"github.com/digitalbitbox/bitbox-wallet-app/util/bolt"
 	"github.com/digitalbitbox/bitbox-wallet-app/util/errp"
+	"github.com/sirupsen/logrus"
 	"go.etcd.io/bbolt"
 )
 
@@ -45,8 +47,8 @@ type DB struct {
 }
 
 // NewDB creates/opens a new db.
-func NewDB(filename string) (*DB, error) {
-	db, err := bbolt.Open(filename, 0600, nil)
+func NewDB(filename string, log *logrus.Entry) (*DB, error) {
+	db, err := bolt.Open(filename, log)
 	if err != nil {
 		return nil, errp.WithStack(err)
 	}