+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Fix a UI bug when checking a backup where the confirmation dialog is sometimes empty
 - Ethereum: remove Ropsten/Rinkeby testnet networks, which have been shut down
 - Re-style Portfolio Chart on mobile (Android) for better usability and a more modern look
+- Gracefully deal with corrupt bbolt database files
 
 ## 4.34.0
 - Bundle BitBox02 firmware version v9.12.0